// apps/web/src/lib/tts/types/openapi/coeiroink.d.ts
export interface paths {
// ...
    "/v1/speakers": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Speakers
         * @description 話者情報の取得。
         */
        get: operations["get_speakers_v1_speakers_get"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v1/synthesis": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Synthesis
         * @description predict+processの処理。
         */
        post: operations["synthesis_v1_synthesis_post"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        Speaker: {
            /** Name */
            name: string;
            /** Speaker Uuid */
            speaker_uuid: string;
            /** Styles */
            styles: components["schemas"]["coeirocore__model__Style"][];
            /** Version */
            version: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}